Add basic `env` config module
